requirements.txt
setuptools
wheel
Kekik
httpx
parsel
fastapi
uvicorn
user_agents
PyYAML
Jinja2
python-multipart
pytest
pytest-asyncio
pytest-mock
motor
redis
pymongo
python-dateutil
numpy
pandas
curl-cffi